test(platform-test): move kubectl and terraform engines into the library
The provisioners need these shell wrappers, and src/ is what `npm run typecheck`
covers, so move the pure mechanics there: kubectl.ts verbatim, terraform.ts
split into a config/fixture-free core plus a thin e2e adapter. The e2e/helpers
entry points become re-export shims so existing test imports keep working. Also
fix two latent e2e type errors (Api vs ApiV4 narrowing).
